[#154] Some fixes
