+++ src/services/supplier_service.ts
@@ -1,8 +1,14 @@
 import { api } from "./api";
-import { POST_SUPPLIER } from "@/constants/api_routes";
+import { POST_SUPPLIER, GET_SUPPLIERS } from "@/constants/api_routes";
 import { useState } from "react";
 import { useGetMarketId } from "@/hooks/useGetMarketId";
 import { dataSupplierInfo } from "@/types/SupplierInfo";
+
+type SupplierService = {
+  getSuppliers: () => Promise<void>;
+  postSupplier: (supplier: dataSupplierInfo) => Promise<boolean> | undefined;
+  suppliers: Supplier;
+}
 
 type Supplier = {
   attributes: {
@@ -16,10 +22,20 @@ type Supplier = {
   type: string;
 }[]
 
-export const SupplierService = () => {
+export const SupplierService = (): SupplierService => {
 
   const { getMarketId } = useGetMarketId()
   const [suppliers, setSuppliers] = useState<Supplier>([]);
+
+  const getSuppliers = async () => {
+    let market_id = await getMarketId()
+    try {
+      const response = await api.get(GET_SUPPLIERS(market_id || ''))
+      setSuppliers(response.data.suppliers.data)
+    } catch (error) {
+      console.log(error)
+    }
+  }
 
   const postSupplier = async (supplier: dataSupplierInfo) => {
     try {
@@ -32,5 +48,5 @@ export const SupplierService = () => {
     }
   }
 
-  return { postSupplier }
+  return { suppliers, getSuppliers, postSupplier } as SupplierService;
 }
